tests: update Plonky3 tests with create_miden_fri_params, 512 row traces

+++ air-script/src/tests/functions/test_air_plonky3.rs
@@ -6,7 +6,7 @@ use crate::{
     tests::functions::functions_complex_plonky3::{FunctionsAir, MAIN_WIDTH},
 };
 
-pub fn generate_trace_rows<F: PrimeField64>(inputs: Vec<u32>) -> RowMajorMatrix<F> {
+pub fn generate_trace_rows<F: PrimeField64>(inputs: Vec<u64>) -> RowMajorMatrix<F> {
     let num_rows = 512;
     let trace_length = num_rows * MAIN_WIDTH;
 
@@ -81,7 +81,7 @@ pub fn generate_trace_rows<F: PrimeField64>(inputs: Vec<u32>) -> RowMajorMatrix<
     trace
 }
 
-fn generate_inputs() -> Vec<u32> {
+fn generate_inputs() -> Vec<u64> {
     vec![1; 16]
 }
